swap: redirty page if page write fails on swap file
Since 62c230b ("mm: add support for a filesystem to activate swap files
and use direct_IO for writing swap pages"), swap_writepage() calls
direct_IO on swap files.  However, in that case the page isn't redirtied
if I/O fails, and is therefore handled afterwards as if it has been
successfully written to the swap file, leading to memory corruption when
the page is eventually swapped back in.

This patch sets the page dirty when direct_IO() fails.  It fixes a memory
corruption that happened while using swap-over-NFS.

Signed-off-by: Jerome Marchand <jmarchan@redhat.com>
Acked-by: Johannes Weiner <hannes@cmpxchg.org>
Acked-by: Mel Gorman <mgorman@suse.de>
Cc: Hugh Dickins <hughd@google.com>
Cc: <stable@vger.kernel.org>	[3.6+]
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>